refactor: refactoring after review
