+++ src/shared/provider/index.ts
@@ -1,1 +1,2 @@
 export { ThemeProvider } from './theme-provider';
+export { AuthProvider, useAuth } from './auth-provider';
